Application is closed after Ctrl+C command instead of requesting a cancellation
